% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{calcll_1d_baf}
\alias{calcll_1d_baf}
\title{calculate baf_1d for 1 normal + 1 tumor case}
\usage{
calcll_1d_baf(lrr, nrc, baf, n_baf, lprior_f_2d, rlprior_f_2d, scale, pscnMax,
  MaxCn)
}
\description{
calculate baf_1d for 1 normal + 1 tumor case
}
\keyword{internal}
