% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{calcll_baf}
\alias{calcll_baf}
\title{calculate baf for 1 normal + 2 tumors case}
\usage{
calcll_baf(lrr, nrc, baf, n_baf, lprior_f_2d, rlprior_f_2d, scale, pscnMax,
  MaxCn)
}
\description{
calculate baf for 1 normal + 2 tumors case
}
\keyword{internal}
