% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{calcll_p1_cpp}
\alias{calcll_p1_cpp}
\title{calculate likelihood for 1 normal + 1 tumor case}
\usage{
calcll_p1_cpp(IT_new, B_new, lp, rlp, var_baf, var_tcn, scale, pscnMax, cnMax)
}
\description{
calculate likelihood for 1 normal + 1 tumor case
}
\keyword{internal}
