% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{calcll_p3}
\alias{calcll_p3}
\title{calculate likelihood for 1 normal + 3 tumors case}
\usage{
calcll_p3(IT_new, B_new, var_baf, var_tcn, scale, pscnMax, cnMax)
}
\description{
calculate likelihood for 1 normal + 3 tumors case
}
\keyword{internal}
