% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/combine_close_seg.R
\name{combine_close_seg}
\alias{combine_close_seg}
\title{combine close segmentation}
\usage{
combine_close_seg(seg_raw, var_baf, data, delta)
}
\description{
combine close segmentation
}
\keyword{internal}
