% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{example_data}
\alias{example_data}
\title{ExampleData}
\format{A data frame with 5 variables: \code{BAF}, \code{chr}, \code{GT}, \code{LRR} and \code{x}, with chromosome number = 22}
\usage{
example_data
}
\description{
ExampleData
}
\keyword{internal}
