% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/preprocessing.R
\name{preprocessing}
\alias{preprocessing}
\title{Preprocessing data}
\usage{
preprocessing(exp_data, normal_snp, tumor_snp, f_path)
}
\arguments{
\item{exp_data}{a string, file name of xxx}

\item{normal_snp}{a string, file name of xxx}

\item{tumor_snp}{a string, file name of xxx}

\item{f_path}{a string, file path of the files above}
}
\value{
\item{df}{a dataframe containing lrr, nrc, baf and n_baf }
}
\description{
Takes \code{exp_data}, \code{normal_snp}, and \code{tumor_snp}
}
\keyword{internal}
