% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{seg_eg}
\alias{seg_eg}
\title{Segmentation}
\format{A data frame with segmentation info, with chromosome = 22}
\usage{
seg_eg
}
\description{
Segmentation
}
\keyword{A}
\keyword{containing...}
\keyword{dataset}
\keyword{internal}
