% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sel_scale_1d.R
\name{sel_scale_1d}
\alias{sel_scale_1d}
\title{select scale for 1 normal + 1 tumor case}
\usage{
sel_scale_1d(cnMax = 6, pscnMax = 6, ngrid = 100, nslaves = 50, B_new,
  IT_new, temp)
}
\description{
select scale for 1 normal + 1 tumor case
}
\keyword{internal}
