% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sel_scale_2d.R
\name{sel_scale_2d}
\alias{sel_scale_2d}
\title{select scale for 1 normal + 2 tumors case}
\usage{
sel_scale_2d(cnMax = 6, pscnMax = 6, ngrid = 100, nslaves = 50, B_new,
  IT_new, temp)
}
\description{
select scale for 1 normal + 2 tumors case
}
\keyword{internal}
