% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sel_scale_2d_wgs.R
\name{sel_scale_2d_wgs}
\alias{sel_scale_2d_wgs}
\title{select scale for 1 normal + 2 tumors case for wgs data}
\usage{
sel_scale_2d_wgs(cnMax = 6, pscnMax = 6, ngrid = 100, nslaves = 50,
  temp)
}
\description{
select scale for 1 normal + 2 tumors case for wgs data
}
\keyword{internal}
